+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 **MapboxCoreSearch**: v2.15.0-beta.1
 
 - [Core] Experimental. Add support for searching parking locations with access to dynamic and static parking metadata.
+- [Core] Remove Carthage integration.
 
 ## 2.14.0